@@ -1,9 +1,9 @@
 {
-    "name": "angular-8-registration-login-example",
+    "name": "angular8-withwebpack",
     "version": "1.0.0",
     "repository": {
         "type": "git",
-        "url": "https://github.com/cornflourblue/angular-8-registration-login-example.git"
+        "url": "https://github.com/cornflourblue/angular8-withwebpack.git"
     },
     "scripts": {
         "build": "webpack --mode production",
@@ -18,6 +18,7 @@
         "@angular/platform-browser": "^8.0.0",
         "@angular/platform-browser-dynamic": "^8.0.0",
         "@angular/router": "^8.0.0",
+        "@types/jest": "^24.0.19",
         "core-js": "^3.1.3",
         "rxjs": "^6.3.3",
         "zone.js": "^0.9.1"
@@ -28,12 +29,12 @@
         "css-loader": "^2.1.1",
         "html-loader": "^0.5.5",
         "html-webpack-plugin": "^3.2.0",
-        "less": "^3.0.4",
-        "less-loader": "^5.0.0",
+        "node-sass": "^4.12.0",
+        "sass-loader": "^8.0.0",
         "style-loader": "^0.23.1",
         "ts-loader": "^6.0.1",
         "typescript": "^3.1.3",
-        "webpack": "^4.32.2",
+        "webpack": "^4.41.2",
         "webpack-cli": "^3.1.2",
         "webpack-dev-server": "^3.7.0"
     }
