@@ -1,24 +1,11 @@
-﻿import { Component } from '@angular/core';
-import { Router } from '@angular/router';
+import { Component ,ViewEncapsulation} from '@angular/core';
+import './app.component.scss';
+// import '../styles/app.scss';
 
-import { AuthenticationService } from './_services';
-import { User } from './_models';
-
-import './_content/app.less';
-
-@Component({ selector: 'app', templateUrl: 'app.component.html' })
+@Component({
+  selector: 'app-root',
+  templateUrl: './app.component.html'
+})
 export class AppComponent {
-    currentUser: User;
-
-    constructor(
-        private router: Router,
-        private authenticationService: AuthenticationService
-    ) {
-        this.authenticationService.currentUser.subscribe(x => this.currentUser = x);
-    }
-
-    logout() {
-        this.authenticationService.logout();
-        this.router.navigate(['/login']);
-    }
+  title = 'Employee records';
 }
