@@ -1,17 +1,13 @@
-﻿import { Routes, RouterModule } from '@angular/router';
+﻿import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { EmployeeRecordsComponent } from './employee-records/employee-records.component';
+import { MenuComponent } from './menu/menu.component';
 
-import { HomeComponent } from './home';
-import { LoginComponent } from './login';
-import { RegisterComponent } from './register';
-import { AuthGuard } from './_helpers';
 
 const routes: Routes = [
-    { path: '', component: HomeComponent, canActivate: [AuthGuard] },
-    { path: 'login', component: LoginComponent },
-    { path: 'register', component: RegisterComponent },
+    {path:'',component: MenuComponent},
+    {path: 'employeerecords', component: EmployeeRecordsComponent}
+  ];
 
-    // otherwise redirect to home
-    { path: '**', redirectTo: '' }
-];
 
 export const appRoutingModule = RouterModule.forRoot(routes);
